Add demo visual window
