+++ src/Main.java
@@ -1,7 +1,8 @@
+import windows.MainWindow;
+
 import javax.swing.*;
 public class Main {
     public static void main(String[] args) {
-
-        System.out.println("hchcha");
+        MainWindow app = new MainWindow();
     }
 }
